pyproject.toml
[project]
name = "chats-expertise"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=5.3.0",
    "expertise-chats",
    "fastapi[standard]>=0.122.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
expertise-chats = { git = "https://github.com/BrenDev0/expertise_chats_shared.git" }
